refactor txn (#119)
* refactor txn

* use moerr
